Improve the markdown parsing
With this commit `omd` and regex is used for the parsing of the markdown
file.

Furthermore, the parsing is improved in the following way: before,
in the changelog, version headers and change headers had to be `##` and
`###` headers respectivelye. Now, they can be of any header type as long
as the type is consistent among all version and change headers
respectively and as long as the version header type is different from
the change header type.